Refactor LinksField logics to reusable trait
